# openapi.yaml
openapi: 3.0.3
info:
  title: BerryFi Studio API
  description: |-
    This document provides the official API specification for the BerryFi Studio Pixel Streaming Platform.
    It is intended for backend developers to implement the required endpoints for the frontend application.
  version: 1.0.0
  contact:
    name: Shubham Bokade
    email: bloggersbokade17@gmail.com
servers:
  - url: https://api.berryfi.studio/v1
    description: Production server for BerryFi Studio API
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Authorization header using the Bearer scheme. The token is a short-lived **access token** obtained from the `/auth/login` or `/auth/refresh` endpoints."

# ============================
# ====== Global Security ======
# ============================
security:
  - BearerAuth: [] # Apply Bearer Authentication globally to all endpoints by default.

# ============================
# ====== API Paths & Endpoints ======
# ============================
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List Workspaces
      description: Get a list of all workspaces accessible to the current user based on their role.
      operationId: listWorkspaces
      responses:
        '200':
          description: List of workspaces.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "apexmarketing"
                        name:
                          type: string
                          example: "Apex Marketing"
                        description:
                          type: string
                          example: "Digital marketing agency specializing in real estate campaigns"
                        adminEmail:
                          type: string
                          example: "david@apexmarketing.com"
                        adminName:
                          type: string
                          example: "Deepak Verma"
                        status:
                          type: string
                          enum: [active, suspended, disabled]
                          example: "active"
                        currentBalance:
                          type: number
                          example: 15000
                        monthlyBudget:
                          type: number
                          example: 3000
                        budgetAction:
                          type: string
                          enum: [alert, block, none]
                          example: "alert"
                        projectId:
                          type: string
                          example: "rav_project"
                        projectName:
                          type: string
                          example: "RAV Project"
                        teamMemberCount:
                          type: integer
                          example: 5
                        sessionsThisMonth:
                          type: integer
                          example: 150
                        creditsUsedThisMonth:
                          type: number
                          example: 1200
                        budgetUsagePercentage:
                          type: number
                          example: 40
                        isOverBudget:
                          type: boolean
                          example: false
                        giftedCredits:
                          type: number
                          example: 5000
                        purchasedCredits:
                          type: number
                          example: 10000
                        totalCreditsUsed:
                          type: number
                          example: 1200
                        remainingGiftedCredits:
                          type: number
                          example: 4000
                        remainingPurchasedCredits:
                          type: number
                          example: 9800
    post:
      tags: [Workspaces]
      summary: Create Workspace
      description: Create a new workspace.
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "New Workspace"
                description:
                  type: string
                  example: "Description of the new workspace"
      responses:
        '201':
          description: Workspace created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
...
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      summary: Get Workspace Details
      description: Get detailed information about a specific workspace.
      operationId: getWorkspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
          example: "apexmarketing"
      responses:
        '200':
          description: Workspace details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
...